requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cramjam==2.6.2
fastparquet==2023.1.0
fsspec==2023.1.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.14
